Splitting the LLM & Prompts api's into separate services
