+++ cli/src/main/java/com/box/l10n/mojito/cli/command/checks/CliCheckerType.java
@@ -11,7 +11,7 @@ public enum CliCheckerType {
   RECOMMEND_STRING_ID_CHECKER(RecommendStringIdChecker.class),
   CONTEXT_COMMENT_REJECT_PATTERN_CHECKER(ContextCommentRejectPatternChecker.class),
   CONTROL_CHARACTER_CHECKER(ControlCharacterChecker.class),
-  OPEN_AI_CHECKER(OpenAIChecker.class);
+  AI_CHECKER(AIChecker.class);
 
   Class<? extends AbstractCliChecker> type;
