+++ webapp/src/main/java/com/box/l10n/mojito/rest/ai/AIException.java
@@ -1,0 +1,12 @@
+package com.box.l10n.mojito.rest.ai;
+
+public class AIException extends RuntimeException {
+
+  public AIException(String message) {
+    super(message);
+  }
+
+  public AIException(String message, Exception e) {
+    super(message, e);
+  }
+}
