+++ webapp/src/main/java/com/box/l10n/mojito/rest/ai/openai/OpenAIConfig.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.rest.openai;
+package com.box.l10n.mojito.rest.ai.openai;
 
 import com.box.l10n.mojito.openai.OpenAIClient;
 import java.util.HashMap;
@@ -10,10 +10,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ConfigurationProperties(prefix = "l10n.openai")
+@ConfigurationProperties(prefix = "l10n.ai.openai")
 public class OpenAIConfig {
-
-  boolean enabled;
 
   String apiKey;
 
@@ -31,10 +29,6 @@ public class OpenAIConfig {
     this.customHeaders = customHeaders;
   }
 
-  public void setEnabled(boolean enabled) {
-    this.enabled = enabled;
-  }
-
   public void setHost(String host) {
     this.host = host;
   }
@@ -44,9 +38,14 @@ public class OpenAIConfig {
   }
 
   @Bean
-  @ConditionalOnProperty(value = "l10n.openai.enabled", havingValue = "true")
+  @ConditionalOnProperty(value = "l10n.ai.service.type", havingValue = "OpenAI")
   public OpenAIClient openAIClient() {
     if (allowRestrictedHeaders != null && !allowRestrictedHeaders.isEmpty()) {
+      /**
+       * Allow modification of a restricted header prior to sending the request.
+       *
+       * <p>Provided list here is a list of header names which are set as unrestricted in the JRE.
+       */
       System.setProperty(
           "jdk.httpclient.allowRestrictedHeaders",
           allowRestrictedHeaders.stream().reduce((a, b) -> a + "," + b).get());
