--- webapp/src/main/java/com/box/l10n/mojito/rest/openai/OpenAIException.java
@@ -1,12 +1,0 @@
-package com.box.l10n.mojito.rest.openai;
-
-public class OpenAIException extends RuntimeException {
-
-  public OpenAIException(String message) {
-    super(message);
-  }
-
-  public OpenAIException(String message, Exception e) {
-    super(message, e);
-  }
-}
