+++ webapp/src/main/java/com/box/l10n/mojito/service/ai/openai/AIPromptRepository.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.service.openai;
+package com.box.l10n.mojito.service.ai.openai;
 
 import com.box.l10n.mojito.entity.AIPrompt;
 import java.util.List;
