+++ webapp/src/test/resources/application-test.properties
@@ -23,5 +23,6 @@ l10n.branchNotification.notifiers.noop.noop-1.enabled=true
 
 l10n.thirdPartySyncJobs.testRepository.thirdPartyProjectId=testProjectId
 
-l10n.openai.enabled=true
-l10n.openai.host=http://somehost.com/
+l10n.ai.enabled=true
+l10n.ai.openai.host=http://somehost.com/
+l10n.ai.service.type=OpenAI
